BUG-99294 cluster creation show cli command fix: mpacks are included
