--- core-api/src/main/java/com/sequenceiq/cloudbreak/api/model/mpack/ClusterResponseMpackDetails.java
@@ -1,22 +1,0 @@
-package com.sequenceiq.cloudbreak.api.model.mpack;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.sequenceiq.cloudbreak.api.model.JsonEntity;
-
-import io.swagger.annotations.ApiModel;
-
-@ApiModel
-@JsonIgnoreProperties(ignoreUnknown = true)
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public class ClusterResponseMpackDetails implements JsonEntity {
-    private String name;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-}
