+++ core/src/main/java/com/sequenceiq/cloudbreak/converter/ManagementPackComponentToClusterResponseMpackDetails.java
@@ -2,15 +2,15 @@ package com.sequenceiq.cloudbreak.converter;
 
 import org.springframework.stereotype.Component;
 
-import com.sequenceiq.cloudbreak.api.model.mpack.ClusterResponseMpackDetails;
+import com.sequenceiq.cloudbreak.api.model.mpack.ManagementPackDetails;
 import com.sequenceiq.cloudbreak.cloud.model.component.ManagementPackComponent;
 
 @Component
 public class ManagementPackComponentToClusterResponseMpackDetails extends
-        AbstractConversionServiceAwareConverter<ManagementPackComponent, ClusterResponseMpackDetails> {
+        AbstractConversionServiceAwareConverter<ManagementPackComponent, ManagementPackDetails> {
     @Override
-    public ClusterResponseMpackDetails convert(ManagementPackComponent source) {
-        ClusterResponseMpackDetails mpack = new ClusterResponseMpackDetails();
+    public ManagementPackDetails convert(ManagementPackComponent source) {
+        ManagementPackDetails mpack = new ManagementPackDetails();
         mpack.setName(source.getName());
         return mpack;
     }
